change state

+++ src/discord/commands/rogue/game-state.mjs
@@ -10,18 +10,15 @@ function loadGameState(id) {
   const saveFile = gameStateFile.replace('[ID]', id)
   if (fs.existsSync(saveFile)) {
     const gameStateString = fs.readFileSync(saveFile)
-    return JSON.parse(gameStateString)
-  } else {
-    fs.writeFileSync(gameStateFile.replace('[ID]', id), '{}')
+    return JSON.parse(gameStateString || '{}')
   }
+
   return null
 }
 
 function deleteGameState(id) {
   const saveFile = gameStateFile.replace('[ID]', id)
-  if (fs.existsSync(saveFile)) {
-    fs.unlinkSync(saveFile)
-  }
+  if (fs.existsSync(saveFile)) fs.unlinkSync(saveFile)
 }
 
 export { saveGameState, loadGameState, deleteGameState }
